refactor: update import paths from _common.js to common/index.js
