+++ packages/torus-sdk-ts/src/main.ts
@@ -7,7 +7,7 @@ import { ApiPromise, WsProvider } from "@polkadot/api";
 import { BasicLogger } from "@torus-network/torus-utils/logger";
 import { formatToken } from "@torus-network/torus-utils/subspace";
 
-import { queryExtFee } from "./chain/_common.js";
+import { queryExtFee } from "./chain/common/index.js";
 import { SS58Address } from "./types/address.js";
 import { sb_string } from "./types/index.js";
